Add files via upload
So... my partner and I never did touch grass : (
But that's okay! Cos we will be touching grass in few minutes. Or at least that's what we've been telling ourselves a couple of times for 42 minutes.

Anyways, just added a little timer (javascript, no styling). When we come back, the world is ours!